Few changes to joindiaspora.com splash page
Hi @maxwell, I hope you don't mind, I've made some suggested changes to the splash page. The main reason is to make it more clear that people can still join Diaspora even though registrations are closed at jd.com - it seems a lot of people are still confused when they visit this page.

Other than that, I've changed 'donate to our project' to 'donate to FSSN', as for tax purposes there's no way to earmark a donation made to FSSN for Diaspora (so this change avoids the IRS claiming that you can), and made a few minor formatting/grammatical changes. I've also tried to make as many of the links as possible Ruby-style links rather than html-style links.

I've also added the asterisk to 'Diaspora' wherever it was missing and made it lower-case (except at the start of a sentence) as this is how the name is branded these days. I've also linked to a lower-case logo for the banner. If you don't like the lower-case logo, you can ditch those changes.

I don't know how to make a PR to the `joindiaspora` branch rather than `develop`, so I hope you don't mind that I've made these changes directly in Github. I've tested the changes locally and they all seem to work ok.

Best, Goob

I hope that's useful!